Major restructuring of the deduplicator
A major restructuring of the deduplicator was performed to clean up the code and allow for easier expansion. A factory methodology has been employed to allow for chaining of various performance monitoring tasks together to not duplicate work (such as performing multiple hash checks/etc). Unit tests were added to help check for issues with features.

The following features were added:

A timeline feature which tracks all transfers and whether or not they were duplicates.

A feature that extracts the destination memory hash to check if the same data is overwriting existing data.